+++ crypto/pbkdf2.h
@@ -43,5 +43,5 @@ private:
 	static CryptoPP::PasswordBasedKeyDerivationFunction* GetPBEwithHmacFunction(HmacScheme hmac_scheme);
 	static size_t GetHmacSize(HmacScheme hmac_scheme);
 public:
-	static vector<byte> derived(const vector<byte>& pwd, const vector<byte>& salt, size_t iterations, HmacScheme hmac_scheme = HmacScheme::HmacSHA1, size_t derived_size = 0);
+	static vector<byte> derived(const vector<byte>& pwd, const vector<byte>& salt, uint32_t iterations, HmacScheme hmac_scheme = HmacScheme::HmacSHA1, size_t derived_size = 0);
 };
